Update to check route and controller
