+++ app/routes/check.routes.js
@@ -5,6 +5,21 @@ const CheckController = require('../controllers/check.controller.js')
 const routes = [
   {
     method: 'POST',
+    path: '/check/flag-for-billing',
+    options: {
+      handler: CheckController.flagForBilling,
+      app: {
+        excludeFromProd: true,
+        plainOutput: true
+      },
+      auth: false,
+      plugins: {
+        crumb: false
+      }
+    }
+  },
+  {
+    method: 'POST',
     path: '/check/licence-return-logs',
     options: {
       handler: CheckController.returnLogsForLicence,
